requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.23
celery==3.1.20
decorator==4.0.9
Django==1.9.4
django-cors-headers==1.1.0
django-filter==0.13.0
django-websocket-redis==0.4.6
djangorestframework==3.3.3
gevent==1.1.0
gnureadline==6.3.3
greenlet==0.4.9
gunicorn==19.1.1
ipython==4.1.2
ipython-genutils==0.1.0
kombu==3.0.35
Markdown==2.6.6
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
pytz==2016.3
PyYAML==3.11
redis==2.10.5
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
